Reworked Variable Parsing.

+++ VisCPU.HL/HLCompilation.cs
@@ -281,14 +281,14 @@ namespace VisCPU.HL
             return VariableMap.ContainsKey( GetFinalName( var ) ) || parent != null && parent.ContainsVariable( var );
         }
 
-        public void CreateVariable( string name, uint size, HLTypeDefinition tdef )
+        public void CreateVariable( string name, uint size, HLTypeDefinition tdef, bool isVisible )
         {
-            VariableMap[GetFinalName( name )] = new VariableData( name, GetFinalName( name ), size, tdef );
+            VariableMap[GetFinalName( name )] = new VariableData( name, GetFinalName( name ), size, tdef, isVisible);
         }
 
-        public void CreateVariable( string name, string content, HLTypeDefinition tdef )
+        public void CreateVariable( string name, string content, HLTypeDefinition tdef, bool isVisible)
         {
-            VariableMap[GetFinalName( name )] = new VariableData( name, GetFinalName( name ), content, tdef );
+            VariableMap[GetFinalName( name )] = new VariableData( name, GetFinalName( name ), content, tdef, isVisible);
         }
 
         public string GetFinalName( string name )
@@ -325,11 +325,13 @@ namespace VisCPU.HL
 
             List < IHLToken > tokens = br.ReadToEnd();
             ParseOneLineStrings( tokens );
+            ParseCharTokens( tokens );
             RemoveComments( tokens );
             ParseImports( tokens );
             ParseIncludes( tokens );
             ParseReservedKeys( tokens, hlpS );
             tokens = tokens.Where( x => x.Type != HLTokenType.OpNewLine ).ToList();
+            ParseVarDefToken( tokens, hlpS );
             ParseBlocks( tokens );
 
             ParseFunctionToken( tokens, hlpS );
@@ -377,25 +379,25 @@ namespace VisCPU.HL
             }
         }
 
-        public void ParseFunctionToken( List < IHLToken > tokens, HLParserSettings settings )
+        public void ParseFunctionToken(List<IHLToken> tokens, HLParserSettings settings)
         {
-            for ( int i = 0; i < tokens.Count; i++ )
+            for (int i = 0; i < tokens.Count; i++)
             {
-                if ( tokens[i].Type == HLTokenType.OpBlockToken )
+                if (tokens[i].Type == HLTokenType.OpBlockToken)
                 {
-                    if ( !HLParsingTools.ReadOneOrNone(
+                    if (!HLParsingTools.ReadOneOrNone(
                                                        tokens,
                                                        i - 1,
                                                        HLTokenType.OpBracketClose,
                                                        out IHLToken bClose
-                                                      ) )
+                                                      ))
                     {
                         continue;
                     }
 
-                    List < IHLToken > argPart = new List < IHLToken > { bClose };
-                    IHLToken[] args = HLParsingTools.ReadUntil( tokens, i - 2, -1, HLTokenType.OpBracketOpen );
-                    argPart.AddRange( args );
+                    List<IHLToken> argPart = new List<IHLToken> { bClose };
+                    IHLToken[] args = HLParsingTools.ReadUntil(tokens, i - 2, -1, HLTokenType.OpBracketOpen);
+                    argPart.AddRange(args);
 
                     IHLToken argOpenBracket = HLParsingTools.ReadOne(
                                                                      tokens,
@@ -403,23 +405,23 @@ namespace VisCPU.HL
                                                                      HLTokenType.OpBracketOpen
                                                                     );
 
-                    argPart.Add( argOpenBracket );
+                    argPart.Add(argOpenBracket);
                     argPart.Reverse();
 
                     int funcIdx = i - 3 - args.Length;
 
-                    if ( tokens[funcIdx].Type != HLTokenType.OpWord )
+                    if (tokens[funcIdx].Type != HLTokenType.OpWord)
                     {
                         continue;
                     }
 
-                    IHLToken funcName = HLParsingTools.ReadOne( tokens, funcIdx, HLTokenType.OpWord );
+                    IHLToken funcName = HLParsingTools.ReadOne(tokens, funcIdx, HLTokenType.OpWord);
 
                     IHLToken typeName = null;
 
-                    if ( funcIdx > 0 &&
-                         ( tokens[funcIdx - 1].Type == HLTokenType.OpWord ||
-                           tokens[funcIdx - 1].Type == HLTokenType.OpTypeVoid ) )
+                    if (funcIdx > 0 &&
+                         (tokens[funcIdx - 1].Type == HLTokenType.OpWord ||
+                           tokens[funcIdx - 1].Type == HLTokenType.OpTypeVoid))
                     {
                         typeName = HLParsingTools.ReadOneOfAny(
                                                                tokens,
@@ -441,14 +443,14 @@ namespace VisCPU.HL
                         int start = modStart - mods.Length + 1;
                         int end = i;
                         IHLToken block = tokens[i];
-                        tokens.RemoveRange( start, end - start + 1 );
+                        tokens.RemoveRange(start, end - start + 1);
 
                         tokens.Insert(
                                       start,
                                       new FunctionDefinitionToken(
                                                                   funcName,
                                                                   typeName,
-                                                                  ParseArgumentList( args.Reverse().ToList() ),
+                                                                  ParseArgumentList(args.Reverse().ToList()),
                                                                   mods,
                                                                   block.GetChildren().ToArray(),
                                                                   start
@@ -457,6 +459,86 @@ namespace VisCPU.HL
 
                         i = start;
                     }
+                }
+            }
+        }
+
+        public void ParseVarDefToken(List<IHLToken> tokens, HLParserSettings settings)
+        {
+            for ( int i = 0; i < tokens.Count; i++ )
+            {
+                if ( tokens[i].Type == HLTokenType.OpWord || settings.MemberModifiers.ContainsValue(tokens[i].Type) )
+                {
+                    List < IHLToken > line = HLParsingTools.ReadUntilAny(
+                                                                      tokens,
+                                                                      i,
+                                                                      1,
+                                                                     new[] { HLTokenType.OpSemicolon , HLTokenType.EOF, HLTokenType.OpBlockBracketOpen, HLTokenType.OpBlockBracketClose}
+                                                                     ).
+                                                            ToList();
+
+                    IHLToken[] mods =
+                        HLParsingTools.ReadNoneOrManyOf( line, 0, 1, settings.MemberModifiers.Values.ToArray() );
+
+                    
+
+                    if ( !HLParsingTools.ReadOneOrNone( line, mods.Length, HLTokenType.OpWord, out IHLToken type ) ||
+                         !HLParsingTools.ReadOneOrNone(line, mods.Length + 1, HLTokenType.OpWord, out IHLToken name))
+                    {
+                        i += line.Count;
+                        continue;
+                    }
+                    IHLToken num = null;
+
+                    if ( line.Count == mods.Length + 2 )
+                    {
+                        tokens.RemoveRange( i, line.Count + 1 );
+                        if (mods.All(x => x.Type != HLTokenType.OpPublicMod) && mods.All(x => x.Type != HLTokenType.OpPrivateMod))
+                        {
+                            mods = mods.Append(new HLTextToken(HLTokenType.OpPrivateMod, "private", 0)).ToArray();
+                        }
+                        tokens.Insert( i, new VariableDefinitionToken( name, type, mods, line.ToArray(), null, null ) );
+
+                        continue;
+                    }
+
+                    if ( line[mods.Length + 2].Type == HLTokenType.OpIndexerBracketOpen )
+                    {
+                        num = HLParsingTools.ReadAny( line, mods.Length + 3 );
+                        HLParsingTools.ReadOne( line, mods.Length + 4, HLTokenType.OpIndexerBracketClose );
+                        tokens.RemoveRange( i, line.Count + 1 );
+                        if (mods.All(x => x.Type != HLTokenType.OpPublicMod) && mods.All(x => x.Type != HLTokenType.OpPrivateMod))
+                        {
+                            mods = mods.Append(new HLTextToken(HLTokenType.OpPrivateMod, "private", 0)).ToArray();
+                        }
+                        tokens.Insert( i, new VariableDefinitionToken( name, type, mods, line.ToArray(), null, num ) );
+
+                        continue;
+                    }
+
+                    if ( line[mods.Length + 2].Type == HLTokenType.OpEquality )
+                    {
+                        tokens.RemoveRange(i, line.Count + 1);
+                        IHLToken[] init = line.Skip( mods.Length + 3 ).ToArray();
+                        if (mods.All(x => x.Type != HLTokenType.OpPublicMod) && mods.All(x => x.Type != HLTokenType.OpPrivateMod))
+                        {
+                            mods = mods.Append(new HLTextToken(HLTokenType.OpPrivateMod, "private", 0)).ToArray();
+                        }
+                        tokens.Insert(
+                                      i,
+                                      new VariableDefinitionToken(
+                                                                  name,
+                                                                  type,
+                                                                  mods,
+                                                                  line.ToArray(),
+                                                                  init,
+                                                                  null
+                                                                 )
+                                     );
+                        continue;
+                    }
+
+                    i += line.Count;
                 }
             }
         }
@@ -492,24 +574,24 @@ namespace VisCPU.HL
             }
         }
 
-        public void ParseOneLineStrings( List < IHLToken > tokens )
+        public void ParseOneLineStrings(List<IHLToken> tokens)
         {
-            for ( int i = 0; i < tokens.Count; i++ )
+            for (int i = 0; i < tokens.Count; i++)
             {
-                if ( i < tokens.Count - 1 &&
-                     tokens[i].Type == HLTokenType.OpDoubleQuote )
+                if (i < tokens.Count - 1 &&
+                     tokens[i].Type == HLTokenType.OpDoubleQuote)
                 {
-                    int idx = tokens.FindIndex( i + 1, t => t.Type == HLTokenType.OpNewLine );
-                    int endQuote = tokens.FindIndex( i + 1, t => t.Type == HLTokenType.OpDoubleQuote );
+                    int idx = tokens.FindIndex(i + 1, t => t.Type == HLTokenType.OpNewLine);
+                    int endQuote = tokens.FindIndex(i + 1, t => t.Type == HLTokenType.OpDoubleQuote);
 
-                    if ( idx == -1 )
+                    if (idx == -1)
                     {
                         idx = tokens.Count - 1;
                     }
 
-                    if ( endQuote == -1 || endQuote > idx )
+                    if (endQuote == -1 || endQuote > idx)
                     {
-                        EventManager < ErrorEvent >.SendEvent(
+                        EventManager<ErrorEvent>.SendEvent(
                                                               new HLTokenReadEvent(
                                                                    HLTokenType.OpDoubleQuote,
                                                                    HLTokenType.OpNewLine
@@ -519,7 +601,7 @@ namespace VisCPU.HL
                         return;
                     }
 
-                    List < IHLToken > content = tokens.GetRange( i + 1, endQuote - i - 1 );
+                    List<IHLToken> content = tokens.GetRange(i + 1, endQuote - i - 1);
 
                     string ConcatContent()
                     {
@@ -536,8 +618,58 @@ namespace VisCPU.HL
                                                         tokens[i].SourceIndex
                                                        );
 
-                    tokens.RemoveRange( i, endQuote - i + 1 );
-                    tokens.Insert( i, newToken );
+                    tokens.RemoveRange(i, endQuote - i + 1);
+                    tokens.Insert(i, newToken);
+                }
+            }
+        }
+
+        public void ParseCharTokens(List<IHLToken> tokens)
+        {
+            for (int i = 0; i < tokens.Count; i++)
+            {
+                if (i < tokens.Count - 1 &&
+                     tokens[i].Type == HLTokenType.OpSingleQuote)
+                {
+                    int idx = tokens.FindIndex(i + 1, t => t.Type == HLTokenType.OpNewLine);
+                    int endQuote = tokens.FindIndex(i + 1, t => t.Type == HLTokenType.OpSingleQuote);
+
+                    if (idx == -1)
+                    {
+                        idx = tokens.Count - 1;
+                    }
+
+                    if (endQuote == -1 || endQuote > idx)
+                    {
+                        EventManager<ErrorEvent>.SendEvent(
+                                                              new HLTokenReadEvent(
+                                                                   HLTokenType.OpSingleQuote,
+                                                                   HLTokenType.OpNewLine
+                                                                  )
+                                                             );
+
+                        return;
+                    }
+
+                    List<IHLToken> content = tokens.GetRange(i + 1, endQuote - i - 1);
+
+                    string ConcatContent()
+                    {
+                        return OriginalText.Substring(
+                                                      content.First().SourceIndex,
+                                                      tokens[i + 1 + content.Count].SourceIndex -
+                                                      content.First().SourceIndex
+                                                     );
+                    }
+
+                    IHLToken newToken = new HLTextToken(
+                                                        HLTokenType.OpCharLiteral,
+                                                        ConcatContent(),
+                                                        tokens[i].SourceIndex
+                                                       );
+
+                    tokens.RemoveRange(i, endQuote - i + 1);
+                    tokens.Insert(i, newToken);
                 }
             }
         }
@@ -640,7 +772,7 @@ namespace VisCPU.HL
                     else
                     {
                         sb.AppendLine(
-                                      $":data {keyValuePair.Value.GetFinalName()} {keyValuePair.Value.Size} linker:hide"
+                                      $":data {keyValuePair.Value.GetFinalName()} {keyValuePair.Value.Size} {( keyValuePair.Value.IsVisible ? "": "linker:hide" )}"
                                      );
                     }
                 }
@@ -792,7 +924,7 @@ namespace VisCPU.HL
                 ProgramCode.Add( $"LOAD {name} {initValue} ;Temp Var House-keeping" );
             }
 
-            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ) );
+            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ), false );
         }
 
         private VariableData GetFreeTempVarCopy( string initValue )
@@ -809,7 +941,7 @@ namespace VisCPU.HL
 
             ProgramCode.Add( $"COPY {initValue} {name} ;Temp Var House-keeping" );
 
-            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ) );
+            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ), false);
         }
 
         private VariableData GetFreeTempVarDref( string initValue )
@@ -826,7 +958,7 @@ namespace VisCPU.HL
 
             ProgramCode.Add( $"DREF {initValue} {name} ;Temp Var House-keeping" );
 
-            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ) );
+            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ), false);
         }
 
         private VariableData GetFreeTempVarLoad( string initValue )
@@ -843,7 +975,7 @@ namespace VisCPU.HL
 
             ProgramCode.Add( $"LOAD {name} {initValue} ;Temp Var House-keeping" );
 
-            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ) );
+            return VariableMap[name] = new VariableData( name, name, 1, TypeSystem.GetOrAdd( "var" ), false);
         }
 
         private string GetPrefix()
